Add test for shutdown script
This commit adds a test for the shutdown script run triggered by the
new exit code introduced earlier. Unfortunately, it's very hard to
do a negative test here because of the way the code is structured
right now and because of `process:exit()`, which would take down the
whole test process if ran.

Signed-off-by: George Pisaltu <georgep@casperlabs.io>